Rename embedding_providers to providers
Rename embedding_providers to providers in tests
