+++ nemoguardrails/embeddings/providers/base.py
