+++ tests/test_kb_openai_embeddings.py
@@ -18,9 +18,7 @@ import os
 import pytest
 
 from nemoguardrails import LLMRails, RailsConfig
-from nemoguardrails.embeddings.embedding_providers.fastembed import (
-    FastEmbedEmbeddingModel,
-)
+from nemoguardrails.embeddings.providers.fastembed import FastEmbedEmbeddingModel
 
 CONFIGS_FOLDER = os.path.join(os.path.dirname(__file__), ".", "test_configs")
